[wip] magit-insert-section: Call magit-maybe-make-margin-overlay here
This handles the case when the section body is inserted at the same
time as the section heading.

`magit--git-wash' already handles the case when the children are
created from Git output, either when the section is inserted or
later when it is expanded.

There might be sections that are not handled by either of those two
approaches.  Those need to keep calling `magit-make-margin-overlay'
explicitly.